+++ docs/whats-new.rst
@@ -10,10 +10,10 @@ What's New
 Coming up next
 --------------
 
-Feature
-^^^^^^^
+Features and front-end API
+^^^^^^^^^^^^^^^^^^^^^^^^^^
 
-- :class:`ArgoIndex` **now support composition of several search criteria**. Thanks to a re-design of the Argo index search engine, it is now easy to use multiple search criteria to query an Argo files index. Checkout the dedicated :ref:`tools-argoindex` documentation page.
+- :class:`ArgoIndex` **now support composition of several search criteria**. Thanks to a re-design of the Argo index search engine, it is now easy to use multiple search criteria to query an Argo files index. Checkout the dedicated :ref:`tools-argoindex` documentation page. (:pr:`470`) by |gmaze|.
 
 Internals
 ^^^^^^^^^